requirements.txt
httpcore==0.15
scikit-build
certifi
onnxruntime
Pillow==8.4.0
opencv-python-headless==4.5.5.64
pyyaml
requests
tinyaes
certifi
requests
werkzeug
paramiko
six
nuitka @ https://github.com/Nuitka/Nuitka/archive/develop.zip
zstandard
pyyaml
easycython
cryptography
requests
loguru
pyOpenSSL
pycryptodome
ordered-set
easycython
cffi
gradio==3.12.0
markdown
fastapi
matplotlib
uvicorn
pytz
jinja2
markupsafe
nuitka
psutil